requirements.txt
streamlit>=1.37.0
openai>=1.40.0
requests>=2.32.3
langchain-openai>=0.1.17
langchain-groq>=0.1.6
langchain-community>=0.2.5
faiss-cpu>=1.8.0
langchain>=0.2.5
tiktoken>=0.7.0
unstructured[pdf]>=0.14.9
google-api-python-client>=2.147.0
serpapi>=0.1.5
gTTS>=2.5.1
pyttsx3>=2.90
SpeechRecognition>=3.10.4
pydub>=0.25.1
matplotlib>=3.9.0
pandas>=2.2.2
python-dotenv>=1.0.1